requirements.txt
pytz==2020.4
python-dateutil==2.8.1
numpy==1.19.4
shapely==1.7.1
fiona==1.8.17
pyproj==2.6.1.post1
geopandas==0.8.1
pandas==1.1.4
sqlalchemy==1.3.20
geoalchemy2==0.8.4
boto3
cantools
cykhash
Cython
daemon
descartes
Fiona
geoalchemy2
geojson
geopandas=0.8.1
geovoronoi
herepy
influxdb
ipykernel
ipython
ipython-genutils
ipywidgets
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
matplotlib
more_itertools
networkx
notebook
numpy
osmium
osmnx
pandas=1.1.4
pickleshare
plotly
polyline
postgis
psycopg2
pyarrow==0.17.1
pygeos
pyrobuf
pyrosm==0.5.2
python-dateutil
pytz
PyYAML
redis
requests
Rtree
scipy
setuptools
Shapely
sphinx
SQLAlchemy
unique-id